refactor: move status list to a file
